+++ src/main/java/ru/murlov/model/ExchangeRate.java
@@ -4,12 +4,12 @@ import java.math.BigDecimal;
 
 public class ExchangeRate {
 
-    private int id;
+    private Long id;
     private final Currency base_currency;
     private final Currency target_currency;
     private final BigDecimal rate;
 
-    public ExchangeRate(int id, Currency base_currency, Currency target_currency, BigDecimal rate) {
+    public ExchangeRate(Long id, Currency base_currency, Currency target_currency, BigDecimal rate) {
         this.id = id;
         this.base_currency = base_currency;
         this.target_currency = target_currency;
@@ -22,7 +22,7 @@ public class ExchangeRate {
         this.rate = rate;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
